wip: refactor route bottom bars code
